Cleaned up StrategicMapKeyboardController (formerly MapMovement) script.
